adding a test and test chracter
