+++ com.unity.perception/Editor/Character.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3ba213bea2d90794694cd2f5a6adce6a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
