+++ com.unity.perception/Editor/Character/CharacterTooling.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 543f5c12a920b31449320b9de408235e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
