+++ com.unity.perception/Editor/Character/CharacterToolingLibrary.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 46e8e04cee4511049ab69310afd5f50f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
