+++ com.unity.perception/Editor/Character/ModelTestUI.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: aa3cd9e043826d8459b8f2c63ed153c6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
