--- com.unity.perception/Editor/Validation.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 3ba213bea2d90794694cd2f5a6adce6a
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
