+++ com.unity.perception/Tests/Editor/CharacterToolingTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 862747af60834f14fa2d28b307a79760
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
