+++ com.unity.perception/Tests/Runtime/TestAssets/Characters.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 87cb21b17fbeb4540ab276ba4ded11a4
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
