+++ com.unity.perception/Tests/Runtime/TestAssets/Characters/sebastian.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 171d5101d9375b04c8154917d7db5df7
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
