+++ com.unity.perception/Tests/Runtime/TestAssets/Characters/sebastian/sebastian_athletic_01_clothes.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2fa8127098d93e44980bfd5270f17eae
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
