+++ com.unity.perception/Tests/Runtime/TestAssets/Characters/sebastian/sebastian_athletic_01_eyes.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f1e19850dd55de54599bf9160641e570
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
