+++ com.unity.perception/Tests/Runtime/TestAssets/Characters/sebastian/sebastian_athletic_01_hair.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 01884d438a534744a9a94d8e4f3905a4
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
