+++ com.unity.perception/Tests/Runtime/TestAssets/Characters/sebastian/sebastian_athletic_01_mouth.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bed28cf0ff190f34db0a7c266ba02f18
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
