+++ com.unity.perception/Tests/Runtime/TestAssets/Characters/sebastian/sebastian_athletic_01_prefab.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: fe79cf9a1b10a5542882420ace5e4b14
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
