+++ com.unity.perception/Tests/Runtime/TestAssets/Characters/sebastian/sebastian_athletic_01_shoes.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 40a435508b00d354d80eef52be4d06cd
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
