+++ com.unity.perception/Tests/Runtime/TestAssets/Characters/sebastian/sebastian_athletic_01_skin.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ff3f4a35617841f40bc13271d471988a
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
